# Use -stdlib=libc++ and -fno-rtti with Xtensa.